% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR_confint.R
\name{IRR_confint}
\alias{IRR_confint}
\title{Calculate confidence interval for incidence rate ratios}
\usage{
IRR_confint(IR_ratio, x1, x2)
}
\arguments{
\item{IR_ratio}{An incidence rate ratio}

\item{x1}{The number of events in the reference group ("unexposed")}

\item{x2}{The number of events in the index group ("exposed")}
}
\value{
A numeric vector of length 2
}
\description{
Calculate confidence interval for incidence rate ratios
}
\examples{
IRR_confint(1.22, 938, 1227)
}
