% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mAssoc.R
\name{mAssoc}
\alias{mAssoc}
\title{Measures of Association}
\usage{
mAssoc(tab, method = "cohort.count", conf.level = 0.95)
}
\arguments{
\item{tab}{table object in epi.2by2() format (referent variable in bottom left, index variable(s) in upper left)}

\item{method}{specification of study type. just like epi.2by2(), options are cohort.count, cohort.time, case.control, or cross.sectional.}

\item{conf.level}{magnitude of the returned confidence intervals. Must be a single number between 0 and 1.}
}
\value{
a list of objects of class epi.2by2 (type ?epi.2by2 for more details)
}
\description{
This is a wrapper function for the {epiR} function epi.2by2(). It allows the user to obtain measures of association in cases where the exposure variable has more than 2 levels.
}
\examples{


}
