% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop.z.test}
\alias{prop.z.test}
\title{Two proportion z-test}
\usage{
prop.z.test(x, n, p = NULL, alternative = c("two.sided", "less", "greater"),
conf.level = 0.95, correct = FALSE, case_col = 1)
}
\arguments{
\item{x}{a vector of length two, providing count
of "successes" or "outcomes of interest" for
either group or a two-dimensional table (or matrix)
with 2 columns, giving the counts of successes and
failures, respectively.}

\item{n}{a vector of length two for counts of trials/observations in either group;
ignored if x is a matrix or a table.}

\item{p}{a vector of length two representing the expected probability
of "success" in either group. The vector's elements must be greater than 0 and less than 1.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default),
"greater" or "less". You can specify just the initial letter. Only used for testing the null that a single proportion equals
a given value, or that two proportions are equal; ignored otherwise.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.
Only used when testing the null that a single proportion equals a given value, or that two proportions are equal; ignored otherwise.}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied where possible.}

\item{case_col}{if x is a table or matrix, a single integer for the index of the column in which cases appear}
}
\description{
Uses z distribution to test for statistical equality of two proportions
}
\details{
Takes a 2x2 table object where default is that cases appear in the first column and counts
of the exposed appear in the second row.

Only groups with finite numbers of successes and failures are used. Counts of
successes and failures must be nonnegative and hence not greater than the corresponding
numbers of trials which must be positive. All finite counts should be integers.

If p is NULL and there is more than one group, the null tested is that the proportions in
each group are the same. If there are two groups, the alternatives are that the probability
of success in the first group is less than, not equal to, or greater than the probability of
success in the second group, as specified by alternative. A confidence interval for the
difference of proportions with confidence level as specified by conf.level and clipped to {-1,1} is returned.
}
