% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_incidence.R
\name{tab_incidence}
\alias{tab_incidence}
\title{Calculate incidence rate and measures of association between strata}
\usage{
tab_incidence(pyears_object, index = 2, ref = 1)
}
\arguments{
\item{pyears_object}{An object generated using survival package function pyears()}

\item{index}{The position of the index group in the vector pyears_object$event}

\item{ref}{The position of the reference group in the vector pyears_object$pyears}
}
\value{
A tibble with calculations for reference group in row 1 and index group in row 2.
}
\description{
Calculate incidence rate and measures of association between strata
}
\examples{
tab_incidence(pyears_male)
}
