% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabodds.R
\name{tabodds}
\alias{tabodds}
\title{Tabulation to display odds}
\usage{
tabodds(
  var_exp,
  var_case,
  data = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)
}
\arguments{
\item{var_exp}{Exposure variable.}

\item{var_case}{Case or outcome variable should be binary vector.}

\item{data}{a data frame object (Optional)}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{print.table}{logical value to display formatted outputs}
}
\description{
\code{tabodds} generates cross-tabulation between two variables and
display odds of failure \code{var_case} against a categorical
explanatory variable \code{var_exp}. It is used with cross-sectional
data.
}
\details{
The variable \code{var_case} should coded 1 for case and 0 for non-case.
A simple table illustrating cases and controls as well as odds for each
category is generated.

\strong{Calculating Odds}

\deqn{OR = d1 x h0 / d0 x h1}

\strong{Error Factor} (EF)

\deqn{EF = exp(1.96 x SE(log odds))}

\deqn{SE(log odds) = \sqrt{1/d + 1/h}}
}
\examples{
\dontrun{

## Asthma Example (page 160, Essential Medical Statistics)
asthma <- expandTables(c(81, 995, 57, 867),
             exp_name = "Sex",
             exp_lvl = c("Women", "Man"),
             case_name = "Asthma",
             case_lvl = c("Yes", "No"))

## labelling
asthma <- labelData(asthma, "Hypothetical Data of Asthma Prevalence")
asthma <- labelVars(asthma,
    c(Sex, Asthma), c("Man or Woman", "Asthma or No Asthma"))

## Checking codebook
codebook(asthma)

## Odds
tabodds(Sex, Asthma, asthma)

## Odds ratios
mhodds(Sex, Asthma, asthma)

## The odds ratio, J Martin Bland, Douglas G Altman, BMJ 2000;320:1468
hay <- expandTables(c(141, 420, 928, 13525),
                    exp_name = "eczema",
                    exp_lvl = c("Yes", "No"),
                    case_name = "hayFever",
                    case_lvl = c("Yes", "No"))
hay <- labelData(hay, "hay fever and eczema in 11 year old children")
hay <- labelVars(hay,
                 c(eczema, hayFever),
                c("prevalence of eczema", "prevalence of hay fever"))


tabodds(eczema, hayFever, hay)
mhodds(eczema, hayFever, hay, "Yes")
}
}
\references{
\enumerate{
\item Essential Medical Statistics, Betty R. Kirwood, Second
Edition
\item Statistics Notes: The odds ratio; J Martin Bland, Douglas G Altman
BMJ 2000;320:1468
}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{descriptive}
\keyword{frequency}
\keyword{odds}
\keyword{odds,}
\keyword{ratio,}
\keyword{statistics,}
\keyword{table,}
